# Conflicts:
#	pxr/imaging/glf/info.cpp
#	pxr/imaging/hdSt/CMakeLists.txt
#	pxr/imaging/hdSt/bufferResource.cpp
#	pxr/imaging/hdSt/bufferResource.h
#	pxr/imaging/hdSt/immediateDrawBatch.cpp
#	pxr/imaging/hdSt/indirectDrawBatch.cpp
#	pxr/imaging/hdSt/renderPassState.cpp
#	pxr/imaging/hdSt/resourceBinder.cpp
#	pxr/imaging/hdSt/volume.cpp
#	pxr/imaging/hdx/drawTargetTask.cpp
#	pxr/imaging/hdx/oitRenderTask.cpp
#	pxr/imaging/hdx/oitResolveTask.cpp
#	pxr/imaging/hdx/pickTask.cpp
#	pxr/imaging/hdx/renderSetupTask.cpp
#	pxr/imaging/hdx/selectionTask.h
#	pxr/imaging/hdx/shadowTask.cpp
#	pxr/imaging/hdx/simpleLightTask.cpp
#	pxr/imaging/hgiInterop/CMakeLists.txt
#	pxr/imaging/hgiInterop/hgiInterop.cpp
#	pxr/imaging/hgiInterop/metal.h
#	pxr/imaging/hgiInterop/metal.mm
#	pxr/usd/plugin/usdMtlx/reader.cpp
#	pxr/usdImaging/usdImagingGL/engine.cpp
#	pxr/usdImaging/usdImagingGL/engine.h

+++ pxr/imaging/hdSt/subdivision.h
@@ -232,7 +232,7 @@ public:
         size_t GetNumElements() const {
             return HdGetComponentCount(_resource->GetTupleType().type);
         }
-        GLuint BindVBO() {
+        uint64_t BindVBO() {
             return _resource->GetId()->GetRawResource();
         }
 #if OPENSUBDIV_HAS_METAL_COMPUTE && defined(PXR_METAL_SUPPORT_ENABLED)
